backend/requirements_flask.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
psutil==5.9.6
langchain==0.1.0
langchain-community==0.0.10
langchain-ollama==0.0.1
langchain-core==0.1.10
faiss-cpu==1.7.4
pypdf==3.17.4

